Fix winner already set for geneation with 2 teams
If a tournament was created with only two teams, the winner was set to the second team on generation of the tournament.
This happened, because the declaration of the winningTeam-variable didn't account for an unset winner and always set the second team as it's value.
This is now prevented by defining it with null initially and using a case statement to initialize it.